Make the path stripping work on hard links
Hard links are special when we strip paths as we need to strip
the target of the hardlink too. In rdup it is only a hardlink
if the link target fails into the set of backed up files - so
using this definition we should be able to shorten the path.