reset the timestamps on the timestamp file
when atime is true, every fill is touched during the
backup. To make rdup not see these files as new in
the backup, we must set the timestamp file with a
timestamp AFTER the backup.
If we do this we will not see file the are changed
DURING the backup... So we only do this when -a is
given and otherwise we restore the