Fir rdup on Aix
I’ve spotted a little bug when using with AIX, when the rdup list is
created, directories have 6 chars for the permissions, so a directory
with 0755 permissions gets written as 200755.  This is a little test of
mine where I create a directory called a with an empty file inside
called a, and do: rdup filelist a > out.  on Aix the contents of out
are:
s breaks rdup-tr as the parsing assumes the file mode will be 4 chars long.
The culprit seems to be the usage of:

st_mode & ~S_IFMT

on AIX this doesn’t seem to be enough on directories (no idea why).  I’ve
replaced all instances of this in the code with:

st_mode & 07777

which clears things up (and works on all platforms).  Patch attached!