Fix tar/pax output when filesize changes
I'm using Ubuntu 11.10 on amd64, and I'm having a problem with filters
in rdup (both 1.0.5 and 1.1.11).

It came to my attention that when using per-file compression the
resulting archive had the same size as a normal tar.
The command (for version 1.1) is:
$ rdup -Pxz -E excl /dev/null /home/ale | rdup-tr -O tar > test.tar

After extracting the tar, xz decompressed the archives while reporting
trailing garbage (which turned to be zeros,
enough to match the original uncompressed file size). Further tests
confirmed this using "-Ptrue" (produces files
full of zeros) and "-Phead,-c,10" (outputs the first 10 bytes and
zeros).